feat: simplify reducers
