+++ src/services/users-setting.ts
@@ -1,5 +1,5 @@
 import axios from "axios"
-import tokenService from './token'
+import tokenService from './user'
 
 const baseUrl = 'http://localhost:3001'
