+++ src/state/reducers/user.ts
@@ -1,0 +1,18 @@
+import { createSlice } from "@reduxjs/toolkit";
+import tokenService from '../../services/user';
+
+const slice = createSlice({
+  name: 'user',
+  initialState: null,
+  reducers: {
+    loginUser(state, { payload }) {
+      return payload
+    },
+    logoutUser(state, { payload }) {
+      return null
+    }
+  },
+})
+
+export const { loginUser, logoutUser } = slice.actions
+export default slice.reducer
